Implement storage trie proof
