+++ prover/src/storage_proof/mod.rs
@@ -1,294 +1,72 @@
-use itertools::Itertools;
-use plonky2::{
-    iop::{
-        target::BoolTarget,
-        witness::{PartialWitness, WitnessWrite},
-    },
-    plonk::{
-        circuit_builder::CircuitBuilder,
-        circuit_data::{CircuitConfig, CircuitData, CommonCircuitData},
-        proof::{ProofWithPublicInputs, ProofWithPublicInputsTarget},
-    },
-    recursion::dummy_circuit::cyclic_base_proof,
-};
-use plonky2_field::types::Field;
-use sp_core::{Blake2Hasher, Hasher};
+use plonky2::plonk::circuit_builder::CircuitBuilder;
 
 use crate::{
     common::{
-        array_to_bits, common_data_for_recursion,
-        targets::{
-            impl_parsable_target_set, impl_target_set, Blake2Target, ParsableTargetSet, TargetSet,
-            VerifierDataTarget,
-        },
-        BuilderExt,
+        targets::{impl_target_set, Blake2Target, TargetSet},
+        ProofComposition,
     },
-    prelude::{consts::BLAKE2_DIGEST_SIZE, *},
-    storage_proof::node_parser::branch_parser::BranchParser,
+    prelude::*,
     ProofWithCircuitData,
 };
-use std::iter;
 
 use self::{
-    hashed_branch_parser::HashedBranchParser, storage_address::PartialStorageAddressTarget,
+    branch_node_chain::BranchNodeChainParserTarget,
+    hashed_leaf_parser::{HashedLeafParser, HashedLeafParserTarget},
+    storage_address::PartialStorageAddressTarget,
 };
 
+mod branch_node_chain;
 mod generic_hasher;
 mod hashed_branch_parser;
 mod hashed_leaf_parser;
 mod node_parser;
 mod storage_address;
 
-const VERIFIER_DATA_NUM_CAP_ELEMENTS: usize = 16;
-
-impl_parsable_target_set! {
-    pub struct BranchNodeChainParserTarget {
-        pub root_hash: Blake2Target,
-        pub leaf_hash: Blake2Target,
-        pub partial_address: PartialStorageAddressTarget,
-    }
-}
+use branch_node_chain::BranchNodeChain;
 
 impl_target_set! {
-    struct BranchNodeChainParserTargetWithVerifierData {
-        root_hash: Blake2Target,
-        leaf_hash: Blake2Target,
-        partial_address: PartialStorageAddressTarget,
-
-        verifier_data: VerifierDataTarget<VERIFIER_DATA_NUM_CAP_ELEMENTS>
+    pub struct StorageProofTarget {
+        pub root_hash: Blake2Target,
+        pub data_hash: Blake2Target,
+        pub address: PartialStorageAddressTarget
     }
 }
 
-pub struct BranchNodeData {
-    pub data: Vec<u8>,
-    pub child_nibble: u8,
+pub struct StorageProof {
+    // TODO: Compose data from node vec.
+    pub branch_node_chain: BranchNodeChain,
+    pub hashed_leaf_parser: HashedLeafParser,
 }
 
-pub struct BranchNodeChain {
-    pub nodes: Vec<BranchNodeData>,
-    pub claimed_leaf_hash: [u8; BLAKE2_DIGEST_SIZE],
-}
+impl StorageProof {
+    pub fn prove(self) -> ProofWithCircuitData<StorageProofTarget> {
+        let branch_node_chain_proof = self.branch_node_chain.prove();
+        let hashed_leaf_parser_proof = self.hashed_leaf_parser.prove();
 
-impl BranchNodeChain {
-    pub fn prove(self) -> ProofWithCircuitData<BranchNodeChainParserTarget> {
-        log::info!("Proving branch node chain...");
+        let composition_builder =
+            ProofComposition::new(branch_node_chain_proof, hashed_leaf_parser_proof);
 
-        let inner = self.inner_proof();
+        let targets_op =
+            |builder: &mut CircuitBuilder<F, D>,
+             branch_node_chain_proof: BranchNodeChainParserTarget,
+             hashed_leaf_parser_proof: HashedLeafParserTarget| {
+                branch_node_chain_proof
+                    .leaf_hash
+                    .connect(&hashed_leaf_parser_proof.node_hash, builder);
 
-        let config = CircuitConfig::standard_recursion_config();
-        let mut builder = CircuitBuilder::new(config);
-        let mut witness = PartialWitness::new();
+                branch_node_chain_proof
+                    .partial_address
+                    .connect(&hashed_leaf_parser_proof.partial_address, builder);
 
-        let public_inputs = builder.recursively_verify_constant_proof(inner, &mut witness);
-
-        BranchNodeChainParserTarget {
-            root_hash: public_inputs.root_hash,
-            leaf_hash: public_inputs.leaf_hash,
-            partial_address: public_inputs.partial_address,
-        }
-        .register_as_public_inputs(&mut builder);
-
-        let result = ProofWithCircuitData::from_builder(builder, witness);
-
-        log::info!("Proven branch node chain");
-
-        result
-    }
-
-    fn inner_proof(self) -> ProofWithCircuitData<BranchNodeChainParserTargetWithVerifierData> {
-        let mut composed_proof: Option<ProofWithCircuitData<_>> = None;
-
-        let mut child_hashes = vec![];
-        for i in 0..self.nodes.len() {
-            if i == self.nodes.len() - 1 {
-                child_hashes.push(self.claimed_leaf_hash);
-            } else {
-                child_hashes.push(Blake2Hasher::hash(&self.nodes[i + 1].data).0);
-            }
-        }
-
-        let root_hash = Blake2Hasher::hash(&self.nodes[0].data).0;
-
-        for (node, claimed_child_hash) in self.nodes.into_iter().zip_eq(child_hashes.into_iter()) {
-            let partial_address_nibbles = if let Some(composed_proof) = &composed_proof {
-                BranchNodeChainParserTarget::parse_public_inputs(
-                    &mut composed_proof.pis().into_iter(),
-                )
-                .partial_address
-                .address
-                .to_vec()
-            } else {
-                vec![]
+                StorageProofTarget {
+                    root_hash: branch_node_chain_proof.root_hash,
+                    data_hash: hashed_leaf_parser_proof.storage_data_hash,
+                    address: hashed_leaf_parser_proof.final_address,
+                }
             };
 
-            let inner_circuit = HashedBranchParser {
-                branch_parser: BranchParser {
-                    node_data: node.data,
-                    claimed_child_node_nibble: node.child_nibble,
-                    partial_address_nibbles,
-                    claimed_child_hash,
-                },
-            };
-
-            let circuit = Circuit::build(inner_circuit);
-
-            let new_proof = if let Some(composed_proof) = composed_proof {
-                circuit.prove_recursive(composed_proof.proof())
-            } else {
-                circuit.prove_initial(root_hash)
-            };
-            composed_proof = Some(new_proof);
-        }
-
-        composed_proof.unwrap()
-    }
-}
-
-struct Circuit {
-    cyclic_circuit_data: CircuitData<F, C, D>,
-
-    common_data: CommonCircuitData<F, D>,
-
-    condition: BoolTarget,
-    inner_cyclic_proof_with_pis: ProofWithPublicInputsTarget<D>,
-
-    witness: PartialWitness<F>,
-}
-
-impl Circuit {
-    fn prove_initial(
-        mut self,
-        root_hash: [u8; BLAKE2_DIGEST_SIZE],
-    ) -> ProofWithCircuitData<BranchNodeChainParserTargetWithVerifierData> {
-        log::info!("    Proving storage trie recursion layer(initial)...");
-
-        let root_hash_bits = array_to_bits(&root_hash);
-        let public_inputs = root_hash_bits
-            .into_iter()
-            .map(|bit| F::from_bool(bit))
-            .enumerate()
-            .collect();
-
-        self.witness.set_bool_target(self.condition, false);
-        self.witness.set_proof_with_pis_target::<C, D>(
-            &self.inner_cyclic_proof_with_pis,
-            &cyclic_base_proof(
-                &self.common_data,
-                &self.cyclic_circuit_data.verifier_only,
-                public_inputs,
-            ),
-        );
-
-        let result =
-            ProofWithCircuitData::from_circuit_data(self.cyclic_circuit_data, self.witness);
-
-        log::info!("    Proven storage trie recursion layer(initial)...");
-
-        result
-    }
-
-    fn prove_recursive(
-        mut self,
-        composed_proof: ProofWithPublicInputs<F, C, D>,
-    ) -> ProofWithCircuitData<BranchNodeChainParserTargetWithVerifierData> {
-        log::info!("    Proving storage trie recursion layer...");
-        self.witness.set_bool_target(self.condition, true);
-        self.witness
-            .set_proof_with_pis_target(&self.inner_cyclic_proof_with_pis, &composed_proof);
-
-        let result =
-            ProofWithCircuitData::from_circuit_data(self.cyclic_circuit_data, self.witness);
-
-        log::info!("    Proven storage trie recursion layer");
-
-        result
-    }
-
-    fn build(inner: HashedBranchParser) -> Circuit {
-        let inner_proof = inner.prove();
-
-        log::info!("    Building storage trie recursion layer...");
-
-        let config = CircuitConfig::standard_recursion_config();
-        let mut builder = CircuitBuilder::new(config);
-        let mut pw = PartialWitness::new();
-
-        let inner_proof_pis = builder.recursively_verify_constant_proof(inner_proof, &mut pw);
-
-        let mut virtual_targets = iter::repeat(()).map(|_| builder.add_virtual_target());
-        let future_inner_cyclic_proof_pis =
-            BranchNodeChainParserTarget::parse(&mut virtual_targets);
-        future_inner_cyclic_proof_pis.register_as_public_inputs(&mut builder);
-
-        let verifier_data_target = builder.add_verifier_data_public_inputs();
-        let common_data = common_data_for_recursion(
-            CircuitConfig::standard_recursion_config(),
-            builder.num_public_inputs(),
-            1 << 13,
-        );
-
-        let condition = builder.add_virtual_bool_target_safe();
-
-        let inner_cyclic_proof_with_pis = builder.add_virtual_proof_with_pis(&common_data);
-        let inner_cyclic_proof_pis = BranchNodeChainParserTargetWithVerifierData::parse_exact(
-            &mut inner_cyclic_proof_with_pis
-                .public_inputs
-                .clone()
-                .into_iter(),
-        );
-        let mut inner_cyclic_proof_pis = BranchNodeChainParserTarget {
-            root_hash: inner_cyclic_proof_pis.root_hash,
-            leaf_hash: inner_cyclic_proof_pis.leaf_hash,
-            partial_address: inner_cyclic_proof_pis.partial_address,
-        };
-
-        inner_cyclic_proof_pis.leaf_hash = builder.select_target_set(
-            condition,
-            &inner_cyclic_proof_pis.leaf_hash,
-            &inner_cyclic_proof_pis.root_hash,
-        );
-        let empty_address = PartialStorageAddressTarget::empty(&mut builder);
-        inner_cyclic_proof_pis.partial_address = builder.select_target_set(
-            condition,
-            &inner_cyclic_proof_pis.partial_address,
-            &empty_address,
-        );
-
-        inner_proof_pis
-            .node_hash
-            .connect(&inner_cyclic_proof_pis.leaf_hash, &mut builder);
-        inner_proof_pis
-            .partial_address
-            .connect(&inner_cyclic_proof_pis.partial_address, &mut builder);
-
-        let final_pis = BranchNodeChainParserTarget {
-            root_hash: inner_cyclic_proof_pis.root_hash,
-            leaf_hash: inner_proof_pis.child_node_hash,
-            partial_address: inner_proof_pis.resulting_partial_address,
-        };
-
-        future_inner_cyclic_proof_pis.connect(&final_pis, &mut builder);
-
-        builder
-            .conditionally_verify_cyclic_proof_or_dummy::<C>(
-                condition,
-                &inner_cyclic_proof_with_pis,
-                &common_data,
-            )
-            .unwrap();
-
-        let cyclic_circuit_data = builder.build::<C>();
-
-        pw.set_verifier_data_target(&verifier_data_target, &cyclic_circuit_data.verifier_only);
-
-        log::info!("    Built storage parser recursion layer");
-
-        Circuit {
-            cyclic_circuit_data,
-            common_data,
-            condition,
-            inner_cyclic_proof_with_pis,
-            witness: pw,
-        }
+        composition_builder
+            .assert_both_circuit_digests()
+            .compose(targets_op)
     }
 }
